+++ hudi-spark-datasource/hudi-spark-common/src/main/scala/org/apache/spark/sql/execution/datasources/parquet/FileGroupReaderBasedParquetFileFormat.scala
@@ -1,46 +1,46 @@
 /*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
  *
- *   http://www.apache.org/licenses/LICENSE-2.0
+ *    http://www.apache.org/licenses/LICENSE-2.0
  *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 
-package org.apache.spark.sql.execution.datasources
+package org.apache.spark.sql.execution.datasources.parquet
 
-import org.apache.hudi.{AvroConversionUtils, HoodieFileIndex, HoodiePartitionCDCFileGroupMapping, HoodiePartitionFileSliceMapping, HoodieTableSchema, SparkAdapterSupport, SparkFileFormatInternalRowReaderContext}
+import org.apache.hudi.{AvroConversionUtils, HoodieFileIndex, HoodiePartitionCDCFileGroupMapping, HoodiePartitionFileSliceMapping, HoodieTableSchema, HoodieTableState, SparkAdapterSupport, SparkFileFormatInternalRowReaderContext}
 import org.apache.hudi.avro.AvroSchemaUtils
 import org.apache.hudi.cdc.{CDCFileGroupIterator, CDCRelation, HoodieCDCFileGroupSplit}
+import org.apache.hudi.client.common.HoodieSparkEngineContext
 import org.apache.hudi.client.utils.SparkInternalSchemaConverter
-import org.apache.hudi.common.config.TypedProperties
+import org.apache.hudi.common.config.{HoodieMemoryConfig, TypedProperties}
 import org.apache.hudi.common.fs.FSUtils
 import org.apache.hudi.common.table.{HoodieTableConfig, HoodieTableMetaClient}
 import org.apache.hudi.common.table.read.HoodieFileGroupReader
+import org.apache.hudi.data.CloseableIteratorListener
 import org.apache.hudi.internal.schema.InternalSchema
+import org.apache.hudi.io.IOUtils
 import org.apache.hudi.storage.StorageConfiguration
 import org.apache.hudi.storage.hadoop.{HadoopStorageConfiguration, HoodieHadoopStorage}
 
 import org.apache.hadoop.conf.Configuration
-import org.apache.hadoop.fs.{FileStatus, Path}
-import org.apache.hadoop.mapreduce.Job
+import org.apache.hadoop.fs.Path
+import org.apache.spark.api.java.JavaSparkContext
 import org.apache.spark.sql.HoodieCatalystExpressionUtils.generateUnsafeProjection
 import org.apache.spark.sql.SparkSession
 import org.apache.spark.sql.catalyst.InternalRow
 import org.apache.spark.sql.catalyst.expressions.JoinedRow
-import org.apache.spark.sql.execution.datasources.HoodieFileGroupReaderBasedFileFormat.{ORC_FILE_EXTENSION, PARQUET_FILE_EXTENSION}
-import org.apache.spark.sql.execution.datasources.orc.OrcFileFormat
-import org.apache.spark.sql.execution.datasources.parquet.{ParquetFileFormat, SparkFileReader}
+import org.apache.spark.sql.execution.datasources.{HoodieFormatTrait, PartitionedFile}
+import org.apache.spark.sql.execution.datasources.FileGroupReaderBasedFileFormat
 import org.apache.spark.sql.execution.vectorized.{OffHeapColumnVector, OnHeapColumnVector}
 import org.apache.spark.sql.internal.SQLConf
 import org.apache.spark.sql.sources.Filter
@@ -50,43 +50,37 @@ import org.apache.spark.util.SerializableConfiguration
 
 import java.io.Closeable
 
-trait HoodieFormatTrait {
-
-  // Used so that the planner only projects once and does not stack overflow
-  var isProjected: Boolean = false
-  def getRequiredFilters: Seq[Filter]
-}
+import scala.collection.JavaConverters.mapAsJavaMapConverter
 
 /**
  * This class utilizes {@link HoodieFileGroupReader} and its related classes to support reading
  * from Parquet formatted base files and their log files.
  */
-class HoodieFileGroupReaderBasedFileFormat(tablePath: String,
-                                           tableSchema: HoodieTableSchema,
-                                           tableName: String,
-                                           queryTimestamp: String,
-                                           mandatoryFields: Seq[String],
-                                           isMOR: Boolean,
-                                           isBootstrap: Boolean,
-                                           isIncremental: Boolean,
-                                           validCommits: String,
-                                           shouldUseRecordPosition: Boolean,
-                                           requiredFilters: Seq[Filter])
-  extends FileFormat with SparkAdapterSupport with HoodieFormatTrait with Serializable {
-  protected var supportBatchCalled = false
-  protected var supportBatchResult = false
+class FileGroupReaderBasedParquetFileFormat(tablePath: String,
+                                            tableSchema: HoodieTableSchema,
+                                            tableName: String,
+                                            queryTimestamp: String,
+                                            mandatoryFields: Seq[String],
+                                            isMOR: Boolean,
+                                            isBootstrap: Boolean,
+                                            isIncremental: Boolean,
+                                            isCDC: Boolean,
+                                            validCommits: String,
+                                            shouldUseRecordPosition: Boolean,
+                                            requiredFilters: Seq[Filter]
+                                                 ) extends ParquetFileFormat with SparkAdapterSupport with HoodieFormatTrait {
 
   def getRequiredFilters: Seq[Filter] = requiredFilters
 
+  private val sanitizedTableName = AvroSchemaUtils.getAvroRecordQualifiedName(tableName)
+
   /**
-   * Conditions to support batch:
-   * 1. Parent file format supports batch,
-   * 2. This read is not for a MOR table,
-   * 3. This read is not an incremental read, and
-   * 4. This read is not a bootstrap read.
-   * TODO: remove query type constraint constraints.
-   * TODO: refactor the logic to be more clear.
+   * Support batch needs to remain consistent, even if one side of a bootstrap merge can support
+   * while the other side can't
    */
+  private var supportBatchCalled = false
+  private var supportBatchResult = false
+
   override def supportBatch(sparkSession: SparkSession, schema: StructType): Boolean = {
     if (!supportBatchCalled || supportBatchResult) {
       supportBatchCalled = true
@@ -95,12 +89,8 @@ class HoodieFileGroupReaderBasedFileFormat(tablePath: String,
     supportBatchResult
   }
 
-  override def isSplitable(sparkSession: SparkSession,
-                           options: Map[String, String],
-                           path: Path): Boolean = false
-
-  // For partition columns that we read from the file, we don't want them to be constant column vectors so we
-  // modify the vector types in this scenario
+  //for partition columns that we read from the file, we don't want them to be constant column vectors so we
+  //modify the vector types in this scenario
   override def vectorTypes(requiredSchema: StructType,
                            partitionSchema: StructType,
                            sqlConf: SQLConf): Option[Seq[String]] = {
@@ -115,8 +105,7 @@ class HoodieFileGroupReaderBasedFileFormat(tablePath: String,
       }
       originalVectorTypes.map {
         o: Seq[String] => o.zipWithIndex.map(a => {
-          if (a._2 >= requiredSchema.length && mandatoryFields.contains(
-            partitionSchema.fields(a._2 - requiredSchema.length).name)) {
+          if (a._2 >= requiredSchema.length && mandatoryFields.contains(partitionSchema.fields(a._2 - requiredSchema.length).name)) {
             regularVectorType
           } else {
             a._1
@@ -126,14 +115,15 @@ class HoodieFileGroupReaderBasedFileFormat(tablePath: String,
     }
   }
 
-  protected val sanitizedTableName = AvroSchemaUtils.getAvroRecordQualifiedName(tableName)
+  private lazy val internalSchemaOpt: org.apache.hudi.common.util.Option[InternalSchema] = if (tableSchema.internalSchema.isEmpty) {
+    org.apache.hudi.common.util.Option.empty()
+  } else {
+    org.apache.hudi.common.util.Option.of(tableSchema.internalSchema.get)
+  }
 
-  protected lazy val internalSchemaOpt: org.apache.hudi.common.util.Option[InternalSchema] =
-    if (tableSchema.internalSchema.isEmpty) {
-      org.apache.hudi.common.util.Option.empty()
-    } else {
-      org.apache.hudi.common.util.Option.of(tableSchema.internalSchema.get)
-    }
+  override def isSplitable(sparkSession: SparkSession,
+                           options: Map[String, String],
+                           path: Path): Boolean = false
 
   override def buildReaderWithPartitionValues(spark: SparkSession,
                                               dataSchema: StructType,
@@ -146,8 +136,7 @@ class HoodieFileGroupReaderBasedFileFormat(tablePath: String,
     val isCount = requiredSchema.isEmpty && !isMOR && !isIncremental
     val augmentedStorageConf = new HadoopStorageConfiguration(hadoopConf).getInline
     setSchemaEvolutionConfigs(augmentedStorageConf)
-    val (remainingPartitionSchemaArr, fixedPartitionIndexesArr) =
-      partitionSchema.fields.toSeq.zipWithIndex.filter(p => !mandatoryFields.contains(p._1.name)).unzip
+    val (remainingPartitionSchemaArr, fixedPartitionIndexesArr) = partitionSchema.fields.toSeq.zipWithIndex.filter(p => !mandatoryFields.contains(p._1.name)).unzip
 
     // The schema of the partition cols we want to append the value instead of reading from the file
     val remainingPartitionSchema = StructType(remainingPartitionSchemaArr)
@@ -159,45 +148,37 @@ class HoodieFileGroupReaderBasedFileFormat(tablePath: String,
     val requestedSchema = StructType(requiredSchema.fields ++ partitionSchema.fields.filter(f => mandatoryFields.contains(f.name)))
     val requestedAvroSchema = AvroConversionUtils.convertStructTypeToAvroSchema(requestedSchema, sanitizedTableName)
     val dataAvroSchema = AvroConversionUtils.convertStructTypeToAvroSchema(dataSchema, sanitizedTableName)
-    val broadcastedParquetFileReader = spark.sparkContext.broadcast(
-      sparkAdapter.createParquetFileReader(supportBatchResult, spark.sessionState.conf, options, augmentedStorageConf.unwrap()))
-    val broadcastedOrcFileReader = spark.sparkContext.broadcast(
-      sparkAdapter.createOrcFileReader(supportBatchCalled, spark.sessionState.conf, options, augmentedStorageConf.unwrap()))
+    val parquetFileReader = spark.sparkContext.broadcast(sparkAdapter.createParquetFileReader(supportBatchResult,
+      spark.sessionState.conf, options, augmentedStorageConf.unwrap()))
     val broadcastedStorageConf = spark.sparkContext.broadcast(new SerializableConfiguration(augmentedStorageConf.unwrap()))
     val fileIndexProps: TypedProperties = HoodieFileIndex.getConfigProperties(spark, options, null)
 
+    val engineContext = new HoodieSparkEngineContext(new JavaSparkContext(spark.sparkContext))
+    val maxMemoryPerCompaction = IOUtils.getMaxMemoryPerCompaction(engineContext.getTaskContextSupplier, options.asJava)
+
     (file: PartitionedFile) => {
       val storageConf = new HadoopStorageConfiguration(broadcastedStorageConf.value.value)
-      file.partitionValues match {
+      val iter = file.partitionValues match {
         // Snapshot or incremental queries.
         case fileSliceMapping: HoodiePartitionFileSliceMapping =>
-          val filegroupName = FSUtils.getFileIdFromFilePath(
-            sparkAdapter.getSparkPartitionedFileUtils.getPathFromPartitionedFile(file))
-          fileSliceMapping.getSlice(filegroupName) match {
+          val fileGroupName = FSUtils.getFileIdFromFilePath(sparkAdapter
+            .getSparkPartitionedFileUtils.getPathFromPartitionedFile(file))
+          fileSliceMapping.getSlice(fileGroupName) match {
             case Some(fileSlice) if !isCount && (requiredSchema.nonEmpty || fileSlice.getLogFiles.findAny().isPresent) =>
-              val fileReaders = new java.util.HashMap[String, SparkFileReader]()
-              fileReaders.put(ORC_FILE_EXTENSION, broadcastedOrcFileReader.value)
-              fileReaders.put(PARQUET_FILE_EXTENSION, broadcastedParquetFileReader.value)
               val metaClient: HoodieTableMetaClient = HoodieTableMetaClient
                 .builder().setConf(storageConf).setBasePath(tablePath).build
-              val readerContext = new SparkFileFormatInternalRowReaderContext(
-                fileReaders, filters, requiredFilters, storageConf, metaClient.getTableConfig)
+              val readers: Map[String, SparkFileReader] = Map("parquet" -> parquetFileReader.value)
+              val readerContext = new SparkFileFormatInternalRowReaderContext(readers.asJava, filters, requiredFilters, storageConf, metaClient.getTableConfig)
               val props = metaClient.getTableConfig.getProps
               options.foreach(kv => props.setProperty(kv._1, kv._2))
-              val reader = new HoodieFileGroupReader[InternalRow](
-                readerContext,
-                new HoodieHadoopStorage(metaClient.getBasePath, storageConf),
-                tablePath,
-                queryTimestamp,
-                fileSlice,
-                dataAvroSchema,
-                requestedAvroSchema,
-                internalSchemaOpt,
-                metaClient,
-                props,
-                file.start,
-                file.length,
-                shouldUseRecordPosition)
+              props.put(HoodieMemoryConfig.MAX_MEMORY_FOR_MERGE.key(), String.valueOf(maxMemoryPerCompaction))
+              val baseFileLength = if (fileSlice.getBaseFile.isPresent) {
+                fileSlice.getBaseFile.get.getFileSize
+              } else {
+                0
+              }
+              val reader = new HoodieFileGroupReader[InternalRow](readerContext, new HoodieHadoopStorage(metaClient.getBasePath, storageConf), tablePath, queryTimestamp,
+                fileSlice, dataAvroSchema, requestedAvroSchema, internalSchemaOpt, metaClient, props, file.start, baseFileLength, shouldUseRecordPosition, false)
               reader.initRecordIterators()
               // Append partition values to rows and project to output schema
               appendPartitionAndProject(
@@ -209,29 +190,18 @@ class HoodieFileGroupReaderBasedFileFormat(tablePath: String,
                 fixedPartitionIndexes)
 
             case _ =>
-              val filePath = sparkAdapter.getSparkPartitionedFileUtils.getPathFromPartitionedFile(file)
-              val baseFileFormat = detectFileFormat(filePath.toString)
-              baseFileFormat match {
-                case PARQUET_FILE_EXTENSION => readBaseFile(file, broadcastedParquetFileReader.value, requestedSchema, remainingPartitionSchema, fixedPartitionIndexes,
-              requiredSchema, partitionSchema, outputSchema, filters, storageConf)
-                case ORC_FILE_EXTENSION => readBaseFile(file, broadcastedOrcFileReader.value, requestedSchema, remainingPartitionSchema, fixedPartitionIndexes,
-                  requiredSchema, partitionSchema, outputSchema, filters, storageConf)
-              }
+              readBaseFile(file, parquetFileReader.value, requestedSchema, remainingPartitionSchema, fixedPartitionIndexes,
+                requiredSchema, partitionSchema, outputSchema, filters, storageConf)
           }
         // CDC queries.
         case hoodiePartitionCDCFileGroupSliceMapping: HoodiePartitionCDCFileGroupMapping =>
-          buildCDCRecordIterator(hoodiePartitionCDCFileGroupSliceMapping, broadcastedParquetFileReader.value, storageConf, fileIndexProps, requiredSchema)
+          buildCDCRecordIterator(hoodiePartitionCDCFileGroupSliceMapping, parquetFileReader.value, storageConf, fileIndexProps, requiredSchema)
 
         case _ =>
-          val filePath = sparkAdapter.getSparkPartitionedFileUtils.getPathFromPartitionedFile(file)
-          val baseFileFormat = detectFileFormat(filePath.toString)
-          baseFileFormat match {
-            case PARQUET_FILE_EXTENSION => readBaseFile(file, broadcastedParquetFileReader.value, requestedSchema, remainingPartitionSchema, fixedPartitionIndexes,
-              requiredSchema, partitionSchema, outputSchema, filters, storageConf)
-            case ORC_FILE_EXTENSION => readBaseFile(file, broadcastedOrcFileReader.value, requestedSchema, remainingPartitionSchema, fixedPartitionIndexes,
-              requiredSchema, partitionSchema, outputSchema, filters, storageConf)
-          }
+          readBaseFile(file, parquetFileReader.value, requestedSchema, remainingPartitionSchema, fixedPartitionIndexes,
+            requiredSchema, partitionSchema, outputSchema, filters, storageConf)
       }
+      CloseableIteratorListener.addListener(iter)
     }
   }
 
@@ -243,10 +213,10 @@ class HoodieFileGroupReaderBasedFileFormat(tablePath: String,
   }
 
   protected def buildCDCRecordIterator(hoodiePartitionCDCFileGroupSliceMapping: HoodiePartitionCDCFileGroupMapping,
-                                     parquetFileReader: SparkFileReader,
-                                     storageConf: StorageConfiguration[Configuration],
-                                     props: TypedProperties,
-                                     requiredSchema: StructType): Iterator[InternalRow] = {
+                                       parquetFileReader: SparkFileReader,
+                                       storageConf: StorageConfiguration[Configuration],
+                                       props: TypedProperties,
+                                       requiredSchema: StructType): Iterator[InternalRow] = {
     val fileSplits = hoodiePartitionCDCFileGroupSliceMapping.getFileSplits().toArray
     val cdcFileGroupSplit: HoodieCDCFileGroupSplit = HoodieCDCFileGroupSplit(fileSplits)
     props.setProperty(HoodieTableConfig.HOODIE_TABLE_NAME_KEY, tableName)
@@ -266,11 +236,11 @@ class HoodieFileGroupReaderBasedFileFormat(tablePath: String,
   }
 
   protected def appendPartitionAndProject(iter: HoodieFileGroupReader.HoodieFileGroupReaderIterator[InternalRow],
-                                        inputSchema: StructType,
-                                        partitionSchema: StructType,
-                                        to: StructType,
-                                        partitionValues: InternalRow,
-                                        fixedPartitionIndexes: Set[Int]): Iterator[InternalRow] = {
+                                          inputSchema: StructType,
+                                          partitionSchema: StructType,
+                                          to: StructType,
+                                          partitionValues: InternalRow,
+                                          fixedPartitionIndexes: Set[Int]): Iterator[InternalRow] = {
     if (partitionSchema.isEmpty) {
       //'inputSchema' and 'to' should be the same so the projection will just be an identity func
       projectSchema(iter, inputSchema, to)
@@ -297,6 +267,7 @@ class HoodieFileGroupReaderBasedFileFormat(tablePath: String,
 
   private def makeCloseableFileGroupMappingRecordIterator(closeableFileGroupRecordIterator: HoodieFileGroupReader.HoodieFileGroupReaderIterator[InternalRow],
                                                           mappingFunction: Function[InternalRow, InternalRow]): Iterator[InternalRow] = {
+    CloseableIteratorListener.addListener(closeableFileGroupRecordIterator)
     new Iterator[InternalRow] with Closeable {
       override def hasNext: Boolean = closeableFileGroupRecordIterator.hasNext
 
@@ -306,10 +277,10 @@ class HoodieFileGroupReaderBasedFileFormat(tablePath: String,
     }
   }
 
-  protected def readBaseFile(file: PartitionedFile, parquetFileReader: SparkFileReader, requestedSchema: StructType,
-                           remainingPartitionSchema: StructType, fixedPartitionIndexes: Set[Int], requiredSchema: StructType,
-                           partitionSchema: StructType, outputSchema: StructType, filters: Seq[Filter],
-                           storageConf: StorageConfiguration[Configuration]): Iterator[InternalRow] = {
+  def readBaseFile(file: PartitionedFile, parquetFileReader: SparkFileReader, requestedSchema: StructType,
+                   remainingPartitionSchema: StructType, fixedPartitionIndexes: Set[Int], requiredSchema: StructType,
+                   partitionSchema: StructType, outputSchema: StructType, filters: Seq[Filter],
+                   storageConf: StorageConfiguration[Configuration]): Iterator[InternalRow] = {
     if (remainingPartitionSchema.fields.length == partitionSchema.fields.length) {
       //none of partition fields are read from the file, so the reader will do the appending for us
       parquetFileReader.read(file, requiredSchema, partitionSchema, internalSchemaOpt, filters, storageConf)
@@ -344,38 +315,4 @@ class HoodieFileGroupReaderBasedFileFormat(tablePath: String,
   private def getFixedPartitionValues(allPartitionValues: InternalRow, partitionSchema: StructType, fixedPartitionIndexes: Set[Int]): InternalRow = {
     InternalRow.fromSeq(allPartitionValues.toSeq(partitionSchema).zipWithIndex.filter(p => fixedPartitionIndexes.contains(p._2)).map(p => p._1))
   }
-
-  override def inferSchema(sparkSession: SparkSession,
-                           options: Map[String, String],
-                           files: Seq[FileStatus]): Option[StructType] = {
-    // This is a simple heuristic assuming all files have the same extension.
-    val fileFormat = detectFileFormat(files.head.getPath.toString)
-
-    val parquetFormat = new ParquetFileFormat()
-    val orcFormat = new OrcFileFormat()
-    fileFormat match {
-      case "parquet" => parquetFormat.inferSchema(sparkSession, options, files)
-      case "orc" => orcFormat.inferSchema(sparkSession, options, files)
-      case _ => throw new UnsupportedOperationException(s"File format $fileFormat is not supported.")
-    }
-  }
-
-  override def prepareWrite(sparkSession: SparkSession,
-                            job: Job,
-                            options: Map[String, String],
-                            dataSchema: StructType): OutputWriterFactory = {
-    throw new UnsupportedOperationException("Write operations are not supported in this example.")
-  }
-
-  private def detectFileFormat(filePath: String): String = {
-    // Logic to detect file format based on the filePath or its content.
-    if (filePath.endsWith(".parquet")) PARQUET_FILE_EXTENSION
-    else if (filePath.endsWith(".orc")) ORC_FILE_EXTENSION
-    else ""
-  }
-}
-
-object HoodieFileGroupReaderBasedFileFormat {
-  val ORC_FILE_EXTENSION = "orc"
-  val PARQUET_FILE_EXTENSION = "parquet"
 }
